+++ core/src/main/java/org/jboss/weld/vertx/AsyncWorker.java
@@ -26,30 +26,30 @@ import io.vertx.core.Vertx;
 import me.escoffier.vertx.completablefuture.VertxCompletableFuture;
 
 /**
- * Allows to perform potentially blocking operations using a Vertx worker thread.
+ * Allows to wrap a synchronous action as an asynchronous computation. The action is performed either as blocking operation using a Vertx worker thread or as
+ * non-blocking operation using the Vertx event-loop thread.
  *
  * <pre>
  * &#64;ApplicationScoped
  * class Hello {
  *
  *     &#64;Inject
- *     BlockingWorker worker;
+ *     AsyncWorker worker;
  *
  *     &#64;Inject
  *     Service service;
  *
  *     CompletionStage&lt;String&gt; hello() {
- *         return worker.perform(() -> service.getNameFromFile()).thenApply((name) -> "Hello " + name + "!");
+ *         return worker.performBlocking(service::getNameFromFile()).thenApply((name) -> "Hello " + name + "!");
  *     }
  *
  * }
  * </pre>
  *
  * @author Martin Kouba
- * @see Vertx#executeBlocking(io.vertx.core.Handler, io.vertx.core.Handler)
  */
 @Dependent
-public class BlockingWorker {
+public class AsyncWorker {
 
     /**
      * Returns a non-contextual instance.
@@ -57,23 +57,43 @@ public class BlockingWorker {
      * @param vertx
      * @return a new worker instance
      */
-    public static BlockingWorker from(Vertx vertx) {
-        return new BlockingWorker(vertx);
+    public static AsyncWorker from(Vertx vertx) {
+        return new AsyncWorker(vertx);
     }
 
     private final Vertx vertx;
 
     @Inject
-    public BlockingWorker(Vertx vertx) {
+    public AsyncWorker(Vertx vertx) {
         this.vertx = vertx;
     }
 
     /**
+     * Performs the specified action using the event-loop thread. The action should never block.
      *
      * @param action
      * @return a completion stage with the result of the specified action
      */
     public <V> CompletionStage<V> perform(Callable<V> action) {
+        VertxCompletableFuture<V> future = new VertxCompletableFuture<>(vertx);
+        vertx.runOnContext(v -> {
+            try {
+                future.complete(action.call());
+            } catch (Throwable e) {
+                future.completeExceptionally(e);
+            }
+        });
+        return future;
+    }
+
+    /**
+     * Performs the specified action using a thread from the worker pool.
+     *
+     * @param action
+     * @return a completion stage with the result of the specified action
+     * @see Vertx#executeBlocking(io.vertx.core.Handler, io.vertx.core.Handler)
+     */
+    public <V> CompletionStage<V> performBlocking(Callable<V> action) {
         VertxCompletableFuture<V> future = new VertxCompletableFuture<>(vertx);
         vertx.<V> executeBlocking((f -> {
             try {
@@ -81,7 +101,7 @@ public class BlockingWorker {
             } catch (Exception e) {
                 f.fail(e);
             }
-        }), (r) -> {
+        }), false, (r) -> {
             if (r.succeeded()) {
                 future.complete(r.result());
             } else {
