--- core/src/test/java/org/jboss/weld/vertx/blocking/Hello.java
@@ -1,39 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2017, Red Hat, Inc., and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.weld.vertx.blocking;
-
-import java.util.concurrent.CompletionStage;
-
-import javax.enterprise.context.Dependent;
-import javax.inject.Inject;
-
-import org.jboss.weld.vertx.BlockingWorker;
-
-@Dependent
-public class Hello {
-
-    @Inject
-    BlockingWorker worker;
-
-    @Inject
-    BlockingFoo foo;
-
-    CompletionStage<String> hello() {
-        return worker.perform(foo::getMessage).thenApply(m -> "Hello " + m + "!");
-    }
-
-}
